+++ src/app/profile/[id]/page.tsx
@@ -1,12 +1,12 @@
 import NotFound from "~/app/not-found"
 import { api } from "~/trpc/server"
 
-import ProfilePage from "./profile/page"
+import ProfilePage from "./profile"
 
 const Profile = async ({ params: { id } }: { params: { id: string } }) => {
   const currentUser = await api.users.getCurrent.query()
 
-  if (currentUser) {
+  if (currentUser.id === id) {
     return <ProfilePage currentUser={currentUser} id={id} />
   }
