+++ src/app/profile/[id]/profile.tsx
